+++ package.json
@@ -16,7 +16,7 @@
     "prepublish": "babel ./src --out-dir ./lib --ignore tests,stories",
     "lint": "eslint src",
     "lintfix": "eslint src --fix",
-    "lint:css": "stylelint './src/**/styles.js'",
+    "lint:css": "stylelint './src/**/styles/*.js'",
     "jest": "jest --no-cache",
     "test:update": "npm run jest -- -u",
     "test": "npm run lint && npm run jest",
